requirements.txt
numpy
cupy-cuda12x; sys_platform == 'linux'
cupy-cuda12x; sys_platform == 'win32'
matplotlib
